Change ReclaimOptions.mSizeThreshold type from uint64_t to int64_t

+++ src/megaapi_impl.cpp
@@ -42365,7 +42365,7 @@ int64_t MegaFileServiceReclaimOptionsPrivate::getReclaimSizeThreshold() const
 
 void MegaFileServiceReclaimOptionsPrivate::setReclaimSizeThreshold(int64_t bytes)
 {
-    mReclaim.mSizeThreshold = static_cast<std::uint64_t>(bytes);
+    mReclaim.mSizeThreshold = bytes;
 }
 
 file_service::ReclaimOptions MegaFileServiceReclaimOptionsPrivate::getOptions() const
